+++ src/components/ParametersTable.vue
