+++ src/components/index.ts
@@ -26,7 +26,6 @@ export { default as DeploymentCombobox } from './DeploymentCombobox.vue'
 export { default as DeploymentDetails } from './DeploymentDetails.vue'
 export { default as DeploymentForm } from './DeploymentForm.vue'
 export { default as DeploymentIconText } from './DeploymentIconText.vue'
-export { default as DeploymentParametersTable } from './DeploymentParametersTable.vue'
 export { default as DeploymentsPageEmptyState } from './DeploymentsPageEmptyState.vue'
 export { default as DeploymentsTable } from './DeploymentsTable.vue'
 export { default as DeploymentToggle } from './DeploymentToggle.vue'
@@ -84,6 +83,7 @@ export { default as PageHeadingWorkQueueCreate } from './PageHeadingWorkQueueCre
 export { default as PageHeadingWorkQueueEdit } from './PageHeadingWorkQueueEdit.vue'
 export { default as PageHeadingWorkQueues } from './PageHeadingWorkQueues.vue'
 export { default as PageHeadingWorkspaceSettings } from './PageHeadingWorkspaceSettings.vue'
+export { default as ParametersTable } from './ParametersTable.vue'
 export { default as RadarApp } from './RadarApp.vue'
 export { default as RadarSmall } from './RadarSmall.vue'
 export { default as ResultsCount } from './ResultsCount.vue'
